utils/pyproject.toml
[project]
name = "ledportal-utils"
version = "0.1.0"
description = "Snapshot processing utilities for LED Portal"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "LED Portal Team" }]
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ledportal_utils"]

[dependency-groups]
dev = ["ruff>=0.4.0", "ty>=0.0.1a6", "pytest>=8.0"]

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM105", # contextlib.suppress - try/except/pass is often clearer
]

[tool.ruff.lint.isort]
known-first-party = ["ledportal_utils"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.environment]
python-version = "3.14"

[tool.ty.rules]
# Strict type checking for utility library
invalid-return-type = "error"
unresolved-import = "error"
unsupported-operator = "warn"
